feat: Se Renderizan las paginas de Inicio y de Cards - se Agrega el NotFound
