+++ src/main.jsx
@@ -1,6 +1,6 @@
 import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
-import App from './pages/App.jsx'
+import App from './routes/App.jsx'
 import './styles/global.scss'
 
 createRoot(document.getElementById('root')).render(
