+++ src/pages/PaginaInicio.jsx
@@ -1,6 +1,7 @@
 // src/componentes/Paginas/PaginaInicio.js
 import React from 'react';
-import MiembroEquipo from '../MiembroEquipo';
+import { Link } from "react-router-dom";
+import MiembroEquipo from '../components/MiembroEquipo';
 
 
 // Ejemplo de datos del equipo
@@ -30,11 +31,19 @@ const equipo = [
 const PaginaInicio = () => {
   return (
     <div className="pagina-inicio">
+      <h1 className="text-center display-1 py-4">
+        Proyecto de Listado y Modificación de Productos
+      </h1>
       <h1>Nuestro Equipo</h1>
       <div className="equipo-lista">
         {equipo.map((miembro, index) => (
           <MiembroEquipo key={index} {...miembro} />
         ))}
+      </div>
+      <div className="text-center mt-4">
+        <Link to="/card">
+          <button className="btn btn-primary">API Rick and Morty</button>
+        </Link>
       </div>
     </div>
   );
